requirements.txt
django
pillow
django-crispy-forms
django-admin-interface
mysqlclient
pymysql
djangorestframework